use SetManager name as more features
